requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
pandas==2.2.2
psycopg2-binary
pyyaml